Add a file for auto-building Mostly I did this as a proof-of-concept---I wanted to come up with a way to automatically build a book without relying on GitHub Pages, which is probably a no-go for proprietary content.
Y'all may find it helpful for editing books yourself---I'd gotten tired of juggling text editor, terminal, and browser myself. If you run this in a terminal, you can just forget about it and switch between editor and browser.
Run with
    python Other/watch.py HRL-Demo-Book
or similar, and saving an edited file will trigger a build.